Further code changes in an effort to address issues raised by PVS-Studio
git-svn-id: https://svn.outpostuniverse.org:8443/svn/outposthd@387 bdf9137d-f3be-7946-88fe-7b0bbfdbf31a

+++ src/UI/Core/ListBox.cpp
@@ -80,7 +80,7 @@ void ListBox::_updateItemDisplay()
 {
 	mItemWidth = rect().width();
 
-	if ((mLineHeight * mItems.size()) > static_cast<int>(height()))
+	if ((mLineHeight * mItems.size()) > static_cast<size_t>(height()))
 	{
 		mLineCount = static_cast<int>(height() / mLineHeight);
 		if (mLineCount < static_cast<int>(mItems.size()))
